# Conflicts:
#	CHANGELOG.md

+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialActorChannel.h
@@ -318,7 +318,7 @@ private:
 	class USpatialReceiver* Receiver;
 
 	FVector LastPositionSinceUpdate;
-	float TimeWhenPositionLastUpdated;
+	double TimeWhenPositionLastUpdated;
 
 	uint8 FramesTillDormancyAllowed = 0;
